+++ indico/modules/events/reminders/notifications_test.py
@@ -8,6 +8,8 @@
 from datetime import datetime, timedelta
 from pathlib import Path
 
+import pytest
+
 from indico.testing.util import assert_email_snapshot
 from indico.web.flask.templating import get_template_module
 
@@ -15,16 +17,24 @@ from indico.web.flask.templating import get_template_module
 pytest_plugins = 'indico.modules.events.timetable.testing.fixtures'
 
 
-def test_event_reminder_email_plaintext(snapshot, create_contribution, create_entry, create_event):
+@pytest.mark.parametrize(('with_label', 'snapshot_filename'), (
+    (False, 'event_reminder.txt'),
+    (True, 'event_reminder_label.txt')
+))
+def test_event_reminder_email_plaintext(snapshot, create_contribution, create_entry, create_event, create_label,
+                                        with_label, snapshot_filename):
     event = create_event(start_dt=datetime(2022, 11, 11, 13, 37),
                          end_dt=datetime(2022, 11, 11, 23, 37), title='Baking with Cats')
     c1 = create_contribution(event, 'Kneading Dough', timedelta(minutes=23))
     c2 = create_contribution(event, 'Pure Bread Cats', timedelta(minutes=40))
     create_entry(c1, datetime(2022, 11, 11, 13, 37))
     create_entry(c2, datetime(2022, 11, 11, 15, 50))
+    if with_label:
+        label = create_label('POSTPONED')
+        event.label = label
     agenda = event.timetable_entries.filter_by(parent_id=None).all()
     template = get_template_module('events/reminders/emails/event_reminder.txt',
                                    event=event, url='http://localhost/', with_description=True,
                                    note='Meow.\nNyah!', with_agenda=True, agenda=agenda)
     snapshot.snapshot_dir = Path(__file__).parent / 'templates/emails/tests'
-    assert_email_snapshot(snapshot, template, 'event_reminder.txt')
+    assert_email_snapshot(snapshot, template, snapshot_filename)